--- lib/msal-react/.prettierignore
@@ -1,1 +1,0 @@
-src/index.ts
