+++ lib/msal-react/src/components/UnauthenticatedTemplate.tsx
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import React, { PropsWithChildren } from "react";
+import { useMsal } from "../hooks/useMsal";
+import { useIsAuthenticated } from "../hooks/useIsAuthenticated";
+import { getChildrenOrFunction } from "../utilities";
+import { AccountIdentifiers } from "../types/AccountIdentifiers";
+
+export type UnauthenticatedTemplateProps = PropsWithChildren<AccountIdentifiers>;
+
+/**
+ * Renders child components if user is unauthenticated
+ * @param props 
+ */
+export function UnauthenticatedTemplate({ username, homeAccountId, children }: UnauthenticatedTemplateProps) {
+    const context = useMsal();
+    const accountIdentifier: AccountIdentifiers = {
+        username,
+        homeAccountId
+    };
+    const isAuthenticated = useIsAuthenticated(accountIdentifier);
+
+    if (!isAuthenticated) {
+        return (
+            <React.Fragment>
+                {getChildrenOrFunction(children, context)}
+            </React.Fragment>
+        );
+    }
+    return null;
+}
