+++ lib/msal-react/src/hooks/useMsal.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { useContext } from "react";
+import { MsalContext } from "../MsalContext";
+
+export const useMsal = () => useContext(MsalContext);
