+++ lib/msal-react/src/types/AccountIdentifiers.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { AccountInfo } from "@azure/msal-browser";
+
+export type AccountIdentifiers = Partial<Pick<AccountInfo, "homeAccountId"|"username">>;
